requirements.txt
pandas==2.1.0
PyYAML==6.0.1
tabulate==0.9.0
PyGithub==2.1.1
requests==2.31.0
